Move TestDatasetDescriptionMap to fuseki-core
