+++ jena-fuseki2/jena-fuseki-webapp/src/test/java/org/apache/jena/fuseki/TS_FusekiWebapp.java
@@ -36,7 +36,6 @@ import org.junit.runners.Suite;
     , TestAdmin.class
     , TestAdminAPI.class
     , TestWebappServerReadOnly.class
-    , TestBuilder.class
     , TestWebappMetrics.class
 })
